+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Changed import of version information to fix bug with importing config.ini
 - Added basic violations/file reporting for some verbosities
-- Refactor of rules to allow rules with memory
+- Refactor of rules to allow rules with memory and added a few memory based rules
 - Refactor of rules to simplify definition
 - Refactor of color cli output to make it more reusable